+++ demo/demo_project/__init__.py
@@ -1,0 +1,4 @@
+try:
+    from demo_project.celery import app as celery_app  # noqa: F401
+except ImportError:
+    pass